pyproject.toml
[tool.poetry]
name = "multimodal-ecommerce-chatbot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
langchain-huggingface = "^0.1.2"
openai-clip = "^1.0.1"
sentence-transformers = "^3.4.1"
langchain-groq = "^0.2.4"
speechrecognition = "^3.14.1"
pyaudio = "^0.2.14"
chainlit = "^2.2.1"
faiss-cpu = "^1.10.0"
torchvision = "^0.21.0"
torch = "^2.6.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
package-mode = false